+++ obsidian/constraints/__init__.py
@@ -1,4 +1,6 @@
 """Constraints: Restrict the recommended space during optimization"""
 
+from .base import *
 from .input import *
 from .output import *
+from .config import *