+++ obsidian/tests/test_constraints.py
@@ -2,10 +2,10 @@
 
 from obsidian.campaign import Campaign
 from obsidian.constraints import (
-    OutConstraint_Blank,
-    InConstraint_Generic,
-    InConstraint_ConstantDim,
-    OutConstraint_L1
+    Linear_Constraint,
+    BatchVariance_Constraint,
+    Blank_Constraint,
+    L1_Constraint
 )
 from obsidian.tests.utils import DEFAULT_MOO_PATH
 
@@ -22,42 +22,52 @@ optimizer = campaign.optimizer
 X_space = campaign.X_space
 target = campaign.target
 
-test_ineq = [[InConstraint_Generic(X_space, indices=[0, 1], coeff=[1, 1], rhs=5)]]
-test_nleq = [[InConstraint_ConstantDim(X_space, dim=0, tol=0.1)]]
-test_out = [[OutConstraint_Blank(target)], [OutConstraint_L1(target, offset=1)]]
+test_linear = [
+    Linear_Constraint(X_space, ind=[0], weights=[1], rhs=5, equality=True),
+    Linear_Constraint(X_space, ind=[0, 1], weights=[1, 1], rhs=5)
+]
+test_nonlinear = [BatchVariance_Constraint(X_space, ind=0, tol=0.1)]
+test_out = [Blank_Constraint(target), L1_Constraint(target, offset=1)]
 
 # Run very short optimizations for testing
 test_config = {'optim_samples': 2, 'optim_restarts': 2}
 
 
-@pytest.mark.parametrize('ineq_constraints', test_ineq)
-def test_ineq_constraints(ineq_constraints):
-    X_suggest, eval_suggest = optimizer.suggest(ineq_constraints=ineq_constraints,
-                                                **test_config)
+@pytest.mark.parametrize('out_const', test_out)
+def test_out_constraints(out_const):
+    out_const.__repr__()
+    campaign.constrain_outputs(out_const)
+    X_suggest, eval_suggest = campaign.optimizer.suggest(**test_config)
     df_suggest = pd.concat([X_suggest, eval_suggest], axis=1)
-    
-    
-@pytest.mark.parametrize('nleq_constraints', test_nleq)
-def test_nleq_constraints(nleq_constraints):
-    X_suggest, eval_suggest = optimizer.suggest(nleq_constraints=nleq_constraints,
-                                                **test_config)
+    campaign.clear_output_constraints()
+
+
+@pytest.mark.parametrize('lin_const', test_linear)
+def test_ineq_constraints(lin_const):
+    lin_const.__repr__()
+    optimizer.X_space.constrain_inputs(lin_const)
+    X_suggest, eval_suggest = optimizer.suggest(**test_config)
     df_suggest = pd.concat([X_suggest, eval_suggest], axis=1)
+    optimizer.X_space.clear_constraints()
     
     
-@pytest.mark.parametrize('out_constraints', test_out)
-def test_out_constraints(out_constraints):
-    X_suggest, eval_suggest = optimizer.suggest(out_constraints=out_constraints,
-                                                **test_config)
+@pytest.mark.parametrize('nl_const', test_nonlinear)
+def test_nleq_constraints(nl_const):
+    nl_const.__repr__()
+    optimizer.X_space.constrain_inputs(nl_const)
+    X_suggest, eval_suggest = optimizer.suggest(**test_config)
     df_suggest = pd.concat([X_suggest, eval_suggest], axis=1)
-    
+    optimizer.X_space.clear_constraints()
+
     
 @pytest.mark.slow
 def test_combo_constraints():
-    X_suggest, eval_suggest = optimizer.suggest(ineq_constraints=test_ineq[0],
-                                                nleq_constraints=test_nleq[0],
+    X_suggest, eval_suggest = optimizer.suggest(ineq_constraints=test_linear[1],
+                                                nleq_constraints=test_nonlinear[0],
                                                 out_constraints=test_out[0],
                                                 **test_config)
     df_suggest = pd.concat([X_suggest, eval_suggest], axis=1)
+    optimizer.X_space.clear_constraints()
     
     
 if __name__ == '__main__':